Move serializable property tests to separate file
